+++ src/MainLib/Marqdouj.DotNet.Web.JsInterop.Azure.Maps/Models/DataDrivenStyles/LinearInterpolateStyleBuilder.cs
@@ -3,7 +3,7 @@
     /// <summary>
     /// Builder class for creating a linear interpolation style expression for Azure Maps.
     /// </summary>
-    internal class LinearInterpolateStyleBuilder
+    public class LinearInterpolateStyleBuilder
     {
         private readonly List<(double Case, string Value)> _cases;